Implement SystemBarIconColorEffect and update TrainerTheme
This commit introduces the `SystemBarIconColorEffect` composable function, which allows dynamic control over the system bar icon color (light or dark) based on the application's theme. It is implemented for both Android and iOS platforms. The `TrainerTheme` composable function is updated to no longer use the `Surface` composable to allow background to be added on the top level in the app. It also does not have the `modifier` parameter anymore.
- Create `SystemBarIconColorEffect.kt` in `commonMain`
- Create `SystemBarIconColorEffect.ios.kt` in `iosMain`
- Create `SystemBarIconColorEffect.android.kt` in `androidMain`
- Update `App` to use the `SystemBarIconColorEffect`
- Remove `Surface` from `TrainerTheme`
- Remove the `modifier` parameter from `TrainerTheme`
- Update the `MainActivity` in the android app to use the new `App` structure.